requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic>=2.9.0
pymongo==4.6.0
requests==2.31.0
email-validator==2.1.0
numpy==1.24.4
opencv-python-headless==4.8.1.78
python-multipart==0.0.6
Pillow==10.0.1
